pyproject.toml
[project]
name="ecs-deplojo"
dynamic = ["version"]
authors = [
    {name = "Lab Digital B.V.", email = "opensource@labdigital.nl"},
]
description="Deployment tool for Amazon ECS"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "boto3>=1.36.2",
    "click>=8.1.8",
    "pyaml>=25.1.0",
    "pytz>=2024.2",
]

[project.scripts]
ecs-deplojo = "ecs_deplojo.cli:main"

[build-system]
requires = ["setuptools", "wheel", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
branch = true
source = ["ecs_deplojo"]

[tool.coverage.paths]
source = ["src", ".tox/*/site-packages"]

[dependency-groups]
test = [
    "coverage[toml]>=7.6.10",
    "moto>=5.0.27",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.2",
]
